Refactor: Graph Schema Entity 파일이름 변경 리팩토링
